[Refactor] 불필요한 파일 삭제 및 코드 가독성 수정
