--- src/Components/Main/MainImageInfo.scss
@@ -1,86 +1,0 @@
-.component_colrabo {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	flex-direction: column;
-	margin: auto;
-	padding: 120px 0 100px 0;
-	background-color: #fafafa;
-
-	.top_text_box {
-		.top_text {
-			margin-bottom: 10px;
-			font-size: 36px;
-			font-weight: 400;
-			letter-spacing: -0.36px;
-		}
-		.top_sign {
-			margin-left: 10px;
-			font-size: 40px;
-			font-weight: 700px;
-		}
-
-		.top_text2 {
-			font-size: 64px;
-			font-weight: 700;
-			line-height: 1.09;
-			letter-spacing: -0.64px;
-		}
-		.line_black_box {
-			display: flex;
-			justify-content: space-between;
-			width: 1140px;
-			margin-top: 59px;
-			height: 12px;
-			.line_black {
-				display: inline-block;
-				position: relative;
-				width: 40%;
-				height: 11px;
-				background-color: black;
-				.line_black2 {
-					position: absolute;
-					right: -25%;
-					display: inline-block;
-					width: 20%;
-					height: 11px;
-					margin-left: 10px;
-					background-color: black;
-				}
-			}
-		}
-	}
-	.image {
-		width: 100vw;
-		margin: 56px 0px 101px;
-	}
-
-	.bottom_text_box {
-		max-width: 1140px;
-		max-height: 211px;
-		.bottom_text {
-			width: inherit;
-			// height: 120px;
-			font-size: 26px;
-			line-height: 40.04px;
-			text-decoration: none;
-		}
-		.botto_common_box {
-			display: flex;
-			justify-content: center;
-			align-items: center;
-			margin-top: 39px;
-			padding: 0px 17px;
-			font-size: 18px;
-			border: 1px solid black;
-			max-width: 160px;
-			height: 52px;
-			.bottom_common_text {
-				text-align: center;
-			}
-			.bottom_common_sign {
-				margin-left: 10px;
-			}
-		}
-	}
-}
