+++ src/Pages/Main/MainImageInfo/MainImageInfo.scss
@@ -12,6 +12,7 @@
 				font-weight: 400;
 				letter-spacing: -0.36px;
 			}
+
 			.top_sign {
 				margin-left: 10px;
 				font-size: 40px;
@@ -25,18 +26,21 @@
 				line-height: 1.09;
 				letter-spacing: -0.64px;
 			}
+
 			.line_black_box {
 				display: flex;
 				justify-content: space-between;
 				max-width: 1140px;
 				margin-top: 59px;
 				height: 12px;
+
 				.line_black {
 					display: inline-block;
 					position: relative;
 					width: 40%;
 					height: 11px;
 					background-color: black;
+
 					.line_black2 {
 						position: absolute;
 						right: -25%;
@@ -64,6 +68,7 @@
 			line-height: 40.04px;
 			text-decoration: none;
 		}
+
 		.bottom_common_box {
 			display: flex;
 			align-items: center;
@@ -71,10 +76,12 @@
 			padding: 0px 30px;
 			border: 1px solid black;
 			height: 52px;
+
 			.bottom_common_text {
 				font-size: 18px;
 				text-align: center;
 			}
+
 			.bottom_common_sign {
 				margin-left: 10px;
 				font-size: 18px;
