+++ src/Pages/ProductDetail/ProductDetail.scss
@@ -82,6 +82,10 @@
 					position: absolute;
 					top: 6px;
 					left: 44px;
+
+					.wish_number {
+						margin-top: 3px;
+					}
 				}
 			}
