+++ src/Pages/ProductList/ProductList.scss
@@ -130,6 +130,10 @@
 					.WishButton {
 						margin: 0;
 						padding: 0;
+
+						.wish_number {
+							margin-top: 2px;
+						}
 					}
 				}
