+++ src/styles/common.scss
@@ -92,6 +92,10 @@ body {
 	letter-spacing: -1%;
 }
 
+.underline {
+	border-bottom: 1px solid #111;
+}
+
 //하트버튼
 .WishButton {
 	display: inline-flex;
@@ -110,10 +114,6 @@ body {
 		font-size: 20px;
 		font-weight: 700;
 		line-height: 21px;
-	}
-
-	.underline {
-		border-bottom: 1px solid #111;
 	}
 }
